requirements.txt
Flask>=1.1,<2.0
flask-mongoengine>=0.9,<1.0
Flask-Cors>=3.0,<4.0
Flask-JWT-Extended>=3.24,<4.0
flask-marshmallow>=0.13,<1.0
marshmallow-mongoengine>=0.11,<1.0
gunicorn>=20.0,<21.0
flask-restx>=0.2,<1.0
python-dotenv>=0.11,<1.0
python-dateutil>=2.8,<3.